LATIR-14: Creacion de componente sobre whatIsLatir
